requirements_growth.txt
# Lean Growth Engine Dependencies
# Add these to your main requirements.txt

# Payment Processing
stripe>=5.0.0

# Analytics
posthog>=3.0.0

# Image Generation
Pillow>=10.0.0

# HTTP Requests (for Airtable/Loops.so)
requests>=2.31.0

# Already required by main app:
# flask
# openai (for voice/photo logging)
